throw if np dl failed

+++ appveyor/install.ps1
@@ -91,14 +91,16 @@ function InstallNumpy() {
         Write-Host $_.Exception
         throw "install numpy failed"
     }
-    mkdir C:\numpy_tmp
-    pushd C:\numpy_tmp
     # download numpy
     #$url="http://www.mirrorservice.org/sites/ftp.sourceforge.net/pub/sourceforge/n/nu/numpy/NumPy/1.9.0/numpy-1.9.0-win32-superpack-python2.7.exe"
     $url="http://www.mi.fu-berlin.de/users/marscher/numpy-1.8.2-cp27-none-win_amd64.whl"
     $webclient = New-Object System.Net.WebClient
     $filepath = "numpy-1.8.2-cp27-none-win_amd64.whl"
     $webclient.DownloadFile($url, $filepath)
+    
+    if (-not Test-Path $filepath) {
+        throw "numpy download failed"
+    }
     # convert installer to wheel
     #RunCommand "python" "-m wheel convert $filepath"
     pip install wheel
